==> notebooks/04.1_dex_proxies_from_daily.ipynb <==
# setup
%load_ext autoreload
%autoreload 2
import os, sys
sys.path.append(os.path.abspath("../src"))

from replication.dex_proxies_from_daily import run_dex_proxies_from_daily, DexProxyTask

"""
# ETH-USDC and ETH-USDT, full window (daily DEX proxies)
run_dex_proxies_from_daily(
    DexProxyTask(
        venues=["uniswap_v2","uniswap_v3"],
        labels=["ETH-USDC","ETH-USDT"]
    )
)
"""

# All pairs, full window
run_dex_proxies_from_daily(
    DexProxyTask(
        venues=["uniswap_v2","uniswap_v3"]
    )
)
