==> notebooks/06_build_daily_comparison_panel.ipynb <==
# setup
%load_ext autoreload
%autoreload 2
import os, sys, datetime as dt
sys.path.append(os.path.abspath("../src"))

from replication.build_daily_comparison_panel import run_build_daily_comparison_panel

"""
# ETH-USDC and ETH-USDT, full window (daily comparison panel)
run_build_daily_comparison_panel(labels=["ETH-USDC","ETH-USDT"])
"""

"""
# All pairs, one week's worth
START = dt.date(2021, 5, 1)
END   = dt.date(2021, 5, 8)
run_build_daily_comparison_panel(start_date=START, end_date=END)
"""

# All pairs, full window
run_build_daily_comparison_panel()
